Change signal_msg[_payload] to vys_signal_msg[_payload]
